Modifications to support PreEditTransform plugins.
The idea is to have plugins that present forms and/or wysiwyg editors
for content.  The default handleEdit behavior is replaced by a plugin
(MyEditArea.hs), and some changes to Interface.hs were made to provide
access to important data for plugins.

This commit includes PreEdit plugins for the default behavior, a
CKEditor edit area plus certain meta-data fields, and
aSPIRE-specific edit_XXX templates to support editing aSPIRE page types.